Add `JPV` instruction
This commit adds the `JPV` operation, which performs a conditional
jump based on the value of the overflow flag in the status register.

This commit also makes all jump insructions (currently `JMP` and `JPV`)
reset the status register (only if a jump was made).

This commit also adds a helper function to make this easier, called
`state_reset_status` (in `src/state.c`).